전투 시스템 리뉴얼(3) + 오류 수정
- URP로 바꾸면서 생긴 타 셰이더 오류 수정
- GridTransparent 셰이더 완성 : 오브젝트의 Scale이 아닌 오직 Cell size에 의존하도록 함
- GetStat의 반환형 float로 수정

+++ Assets/GamePlay/Common/Scripts/Entities/Character/CharacterUtil/DamageCalculator.cs
@@ -16,12 +16,12 @@ public static class DamageCalculator
         List<SystemEnum.eKeyword> inputKeywords, 
         CharBase client)
     {
-        List<float> statInputs = inputStats.Select(stat => client.CharStat.GetStat(stat)).ToList();
+        List<long> statInputs = inputStats.Select(stat => client.CharStat.GetStat(stat)).ToList();
         List<float> keywordInputs = inputKeywords.Select(kw => (float)client.KeywordInfo.GetKeywordCount(kw)).ToList();
         return Evaluate(formula, statInputs, keywordInputs);
     }
 
-    public static float Evaluate(string formula, List<float> inputStats, List<float> inputKeywords)
+    public static float Evaluate(string formula, List<long> inputStats, List<float> inputKeywords)
     {
         if(string.IsNullOrWhiteSpace(formula)) return 0f;
