+++ Assets/GamePlay/Common/Scripts/Entities/Character/Components/CharStat.cs
@@ -92,7 +92,7 @@ namespace AngelBeat
         }
         #endregion
         
-        public float GetStat(eStats stat)
+        public long GetStat(eStats stat)
         {
             return _charStat[(int)GetProperStatAttribute(stat)];
         }
